Moving the recipe construction to seperate function
